Feature/t68 refactorización general de proyecto (#80)
* se migra CreateCommunityViewModel.swift a la carpeta de viewModels, y se elimina la carpeta de viewmodel

* Se migra CreateCommunityViewModel a la carpeta de viewmodels, y se limina la carpeta viewmodel

* Se han migrado correctamente los modelos de datos a estructuras más claras, Todas las vistas han sido refactorizadas para depender de ViewModels.

* Se migra las funciones profileImage() y selectAvatarView al archivo NewAccountViewModel, y se añades algunos comentarios

* se eliminan fragmento de codigo comentado que no es util

* Se elimina archivos inecesarios tras la migracion ChatsServices, ChatsViewModel
