+++ Bubble/Preview Content/DataListMock.swift
@@ -7,20 +7,20 @@
 
 import Foundation
 
-struct ModelListMock: Identifiable {
-    var id = UUID()
-    var nameAlias: String
-    var nameImage: String
-    var dataTimer: String
-}
-
-@MainActor let dataList: [ModelListMock] = [
-  .init(nameAlias: "TopGum", nameImage: "topgum", dataTimer: "01:50"),
-    .init(nameAlias: "Peterete", nameImage: "peterete", dataTimer: "18:20"),
-    .init(nameAlias: "Paloma23", nameImage: "paloma", dataTimer: "16:30"),
-    .init(nameAlias: "Yeikobu", nameImage: "yey", dataTimer: "09:30"),
-    .init(nameAlias: "Lordzzz", nameImage: "lordzzz", dataTimer: "10:15")/*,
-    .init(nameAlias: "Lovezno", nameImage: "lovezno", dataTimer: "11:00"),
-    .init(nameAlias: "Monica36", nameImage: "monica", dataTimer: "20:07"),
-    .init(nameAlias: "Veronica27", nameImage: "veronica", dataTimer: "21:07")*/
-]
+//struct ModelListMock: Identifiable {
+//    var id = UUID()
+//    var nameAlias: String
+//    var nameImage: String
+//    var dataTimer: String
+//}
+//
+//@MainActor let dataList: [ModelListMock] = [
+//  .init(nameAlias: "TopGum", nameImage: "topgum", dataTimer: "01:50"),
+//    .init(nameAlias: "Peterete", nameImage: "peterete", dataTimer: "18:20"),
+//    .init(nameAlias: "Paloma23", nameImage: "paloma", dataTimer: "16:30"),
+//    .init(nameAlias: "Yeikobu", nameImage: "yey", dataTimer: "09:30"),
+//    .init(nameAlias: "Lordzzz", nameImage: "lordzzz", dataTimer: "10:15")/*,
+//    .init(nameAlias: "Lovezno", nameImage: "lovezno", dataTimer: "11:00"),
+//    .init(nameAlias: "Monica36", nameImage: "monica", dataTimer: "20:07"),
+//    .init(nameAlias: "Veronica27", nameImage: "veronica", dataTimer: "21:07")*/
+//]
